+++ src/py42/clients/securitydata.py
@@ -172,18 +172,37 @@ class SecurityDataClient(object):
         )
 
     def search_file_events(self, query):
-        """Searches for file events.
+        """Searches for file events, returns up to the first 10,000 events.
         `REST Documentation <https://support.code42.com/Administrator/Cloud/Monitoring_and_managing/Forensic_File_Search_API>`__
 
         Args:
-            query (:class:`py42.sdk.queries.fileevents.file_event_query.FileEventQuery`): Also
-                accepts a raw JSON str.
+            query query (`:class:`py42.sdk.queries.fileevents.file_event_query.FileEventQuery` or str): The file event query to filter search results.
 
         Returns:
             :class:`py42.response.Py42Response`: A response containing the first 10,000
             events.
         """
         return self._file_event_service.search(query)
+
+    def search_all_file_events(self, query, page_token=""):
+        """Searches for all file events, returning a page of events with a token in the response to retrieve next page.
+
+        `REST Documentation <https://support.code42.com/Administrator/Cloud/Monitoring_and_managing/Forensic_File_Search_API>`__
+
+        Args:
+            query query (`:class:`py42.sdk.queries.fileevents.file_event_query.FileEventQuery` or str): The file event query
+                to filter search results.
+            page_token (str, optional): A token used to indicate the starting point for additional page results. Defaults to empty string.
+                For the first page, do not pass page_token. For all consecutive pages, pass the token from the previous response
+                from field `nextPgToken`.
+
+        Returns:
+            :class:`py42.response.Py42Response`: A response containing page of events.
+        """
+
+        query.page_token = page_token
+        response = self._file_event_service.search(query)
+        return response
 
     def stream_file_by_sha256(self, checksum):
         """Stream file based on SHA256 checksum.